If the current page is in the navigation, do not link to it
When using the `includeInBreadcrumbs` option, the current page gets shown as the last item in the breadcrumb.

Currently it shows as a link to itself and there is no indication that it is the current page.

In this change, if the current page is in the navigation, no href is included, so the page does not get linked to by the breadcrumb component.

This creates 2 options for breadcrumbs:

1. Hide the current page, only show the parents (current design system recommendation)
2. Show the current page, and indicate that it's the current page by not making it a link (previous design system recommendation)