Set opengraphImageUrl in parent tempalte
This changes the code so that `opengraphImageUrl` is always set, either to the specific image for the page, or to the generic image in settings.

By setting this for the parent template to include, `<meta property="og:image">` is no longer needed within this template (but we still need the Twitter variant).