refactor: allow missing nav items
When the "eleventyExcludeFromCollections" option is enabled, the
"collections.all" array will not contain all pages, resulting in
eleventyNavigationBreadcrumb() throwing a "Node does not exist"
exception, which manifests as an opaque nunjucks error that is difficult
for the reader to trace.

This change uses the "allowMissing" option of eleventyNavigationBreadcrumb()
to effectively allow this scenario to fail silently (no breadcrumb will
appear at all rather than failing to compile).